requirements.txt
fastapi
uvicorn
sqlalchemy
xlwings
psd-tools
Pillow
taskiq
taskiq-fastapi
openpyxl